fix(recipes): bound conditional schema evaluation
