+++ .github/file-size-exemptions.txt
@@ -17,7 +17,3 @@ crates/dcc-mcp-gateway/src/gateway/admin/trace.rs
 crates/dcc-mcp-skill-rest/src/service.rs
 # PIP-1799 — capability_service.rs 1507L, exceeds 1500 limit by 7 lines
 crates/dcc-mcp-gateway/src/gateway/capability_service.rs
-# #2385 — owner: Core maintainers; recipes.py owns the dependency-free Draft
-# 2020-12 validator; remove after the dedicated schema-module refactor lands.
-# Review deadline: next Core schema maintenance cycle.
-python/dcc_mcp_core/recipes.py
